Update Readme
This update introduces a guide on creating a custom `ResourceBrowser` component and configuring the plugin to use it. Additionally, details about available render hooks and how to register them are also included. This enhancement can be particularly useful when adding extra functions such as support for the Advanced Table plugin.